fix: HospitalService 테스트 위한 조정, 도메인 담당자 커밋 후 삭제할 것
