+++ src/main/java/com/dentallink/domain/bookmark/service/FavoriteServiceImpl.java
@@ -3,7 +3,7 @@ package com.dentallink.domain.bookmark.service;
 import com.dentallink.domain.bookmark.entity.Favorite;
 import com.dentallink.domain.bookmark.enums.FavoriteAction;
 import com.dentallink.domain.bookmark.repository.FavoriteRepository;
-import com.dentallink.domain.hospital.service.HospitalService;
+import com.dentallink.domain.hospital.service.HospitalExternalService;
 import com.dentallink.domain.user.dto.security.AuthUser;
 import com.dentallink.domain.user.service.query.UserQueryService;
 import lombok.RequiredArgsConstructor;
@@ -18,7 +18,7 @@ public class FavoriteServiceImpl implements FavoriteService {
 
     private final FavoriteRepository favoriteRepository;
     private final UserQueryService userQueryService;
-    private final HospitalService hospitalService;
+    private final HospitalExternalService hospitalService;
 
     @Transactional
     @Override
