+++ src/main/java/com/dentallink/domain/hospital/service/HospitalExternalService.java
@@ -8,6 +8,7 @@ import com.dentallink.domain.hospital.dto.response.*;
 import com.dentallink.domain.hospital.entity.Hospital;
 import com.dentallink.domain.hospital.entity.HospitalSchedule;
 import com.dentallink.domain.hospital.exception.HospitalErrorCode;
+import com.dentallink.domain.hospital.repository.HospitalRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class HospitalExternalService {
     private final HospitalInternalService hospitalInternalService;
+    private final HospitalRepository hospitalRepository;
 
     // 병원 등록
     @Transactional
@@ -120,5 +122,11 @@ public class HospitalExternalService {
         if (!hospital.getUserId().equals(userId)) {
             throw new GlobalException(HospitalErrorCode.NOT_HOSPITAL_OWNER);
         }
+    }
+
+    @Transactional
+    public Hospital getHospitalById(Long id) {
+        return hospitalRepository.findById(id)
+                .orElseThrow(() -> new GlobalException(HospitalErrorCode.HOSPITAL_NOT_FOUND));
     }
 }
